+++ src/AWS.ts
@@ -63,7 +63,8 @@ import { PublicRole } from './provider/aws/iam/roles/cmds/publicRole'
 import { HasManagedAdmin as users_HasManagedAdmin } from './provider/aws/iam/users/cmds/hasManagedAdmin'
 import { MaxKeyAge } from './provider/aws/iam/users/cmds/maxKeyAge'
 import { PasswordAge } from './provider/aws/iam/users/cmds/passwordAge'
-import { RootUserMfaEnabled } from './provider/aws/iam/users/cmds/rootUserMfaEnabled'
+import { RootUser } from './provider/aws/iam/users/cmds/rootUser'
+import { UserMfaEnabled } from './provider/aws/iam/users/cmds/userMfaEnabled'
 export const iam = {
   groups: {
     HasManagedAdmin: groups_HasManagedAdmin,
@@ -76,7 +77,8 @@ export const iam = {
     HasManagedAdmin: users_HasManagedAdmin,
     MaxKeyAge,
     PasswordAge,
-    RootUserMfaEnabled,
+    RootUser,
+    UserMfaEnabled,
   },
 }
 
@@ -119,6 +121,8 @@ export const sqs = {
 }
 
 import { EmrCluster } from './provider/aws/emr/cmds/emrCluster'
+import { SecurityConfigSetting } from './provider/aws/emr/cmds/securityConfigSetting'
 export const emr = {
   EmrCluster,
+  SecurityConfigSetting,
 }