+++ src/provider/aws.ts
@@ -36,7 +36,7 @@ const epilogue = `
 
   This also means that when looking for resources that have admin more then one
   scan is necessary.  For example when looking for users with admin it is
-  necessary to check the user, the user's, groups, and any inline policy they
+  necessary to check the user, the user's groups, and any inline policy they
   have.
 
 ********************************************************************************
@@ -63,6 +63,7 @@ export const builder = (yargs: Argv) => {
     .commandDir('./aws/secrets-manager', opts)
     .commandDir('./aws/emr', opts)
     .commandDir('./aws/efs', opts)
+    .commandDir('./aws/cloudwatch', opts)
     .demandCommand(1)
     .option('region', {
       alias: 'r',