+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- EFS - EncryptionEnabled
+- Cloudwatch Logs - EncryptionEnabled
+
 - AWS - EFS - EncryptionEnabled
 
 ## [1.11.0] - 2020-11-28